+++ src/passes/PrintCallGraph.cpp
@@ -96,12 +96,10 @@ struct PrintCallGraph : public Pass {
     CallPrinter printer(module);
 
     // Indirect Targets
-    for (auto& table : module->tables) {
-      for (auto& segment : table->segments) {
-        for (auto& curr : segment.data) {
-          auto* func = module->getFunction(curr);
-          o << "  \"" << func->name << "\" [style=\"filled, rounded\"];\n";
-        }
+    for (auto& segment : module->elementSegments) {
+      for (auto& curr : segment->data) {
+        auto* func = module->getFunction(curr);
+        o << "  \"" << func->name << "\" [style=\"filled, rounded\"];\n";
       }
     }