+++ src/passes/RemoveImports.cpp
@@ -49,11 +49,9 @@ struct RemoveImports : public WalkerPass<PostWalker<RemoveImports>> {
       *curr, [&](Function* func) { names.push_back(func->name); });
     // Do not remove names referenced in a table
     std::set<Name> indirectNames;
-    for (auto& table : curr->tables) {
-      for (auto& segment : table->segments) {
-        for (auto& name : segment.data) {
-          indirectNames.insert(name);
-        }
+    for (auto& segment : curr->elementSegments) {
+      for (auto& name : segment->data) {
+        indirectNames.insert(name);
       }
     }
     for (auto& name : names) {